Huge Highlight Refactor
Decided to refactor nearly the entirety of how the highlight system works. Instead of having a whole bunch of "ranges" for each highlighted line that goes from the start to the end of the highlighted text on each line, instead the position at which the cursor was when the user started highlighting is cached and everything needed (so far) can be calculated from there (this is what the helper function GetHighlightInfo does). May need to shorten name of "highlightInfo" but dk what other name suffices.