Removed Config Code Generator
Realised that generating a struct from the config.txt wasn't the best idea so just decided to get rid of the idea and instead just have introspection on the struct to generate MemberMetaData values for the struct. Will need some more work once we start introspecting more than one struct but suffices for now.