Add includes directory to repo + Gifs Added as Files
Added the includes directory to the repo so it should be a bit easier for others to clone the repo and get going. Will later add the fonts repo once I figure out licensing stuff.

Added the README gifs as files so that they can be higher quality (fingers crossed).